+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # CHANGELOG
 
 ## [Unreleased]
+
+### Added
+- Inspector support: `withInspector()` enables `PWDEBUG` on the Node server and `Page::pause()` opens Playwright Inspector.